Move CheckRequiredVersion() test to correct source file
